extract some more
